+++ resources/lang/ru.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
+use Serhii\Ago\Lang;
 use Serhii\Ago\LangForm;
 use Serhii\Ago\LangSet;
 
 return new LangSet(
-    lang: "ru",
+    lang: Lang::RU,
     format: "{num} {timeUnit} {ago}",
     ago: "назад",
     online: "В сети",