+++ src/LangForm.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Serhii\Ago;
 
-final readonly class LangForm
+final class LangForm
 {
     public function __construct(
-        public string $other,
-        public string|null $zero = null,
-        public string|null $one = null,
-        public string|null $few = null,
-        public string|null $many = null,
+        public readonly string $other,
+        public readonly string|null $zero = null,
+        public readonly string|null $one = null,
+        public readonly string|null $few = null,
+        public readonly string|null $many = null,
     ) {
     }
 }